requirements.txt
Flask==2.3.3
flask-login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
requests==2.31.0
transformers==4.38.0
torch>=2.0.0
flask-mail==0.9.1
gunicorn==21.2.0
pandas==2.2.2
Werkzeug==2.3.7
email-validator==2.1.0
feedparser==6.0.10
authlib==1.6.1
requests-oauthlib==2.0.0